refactor(arel) extract ArelConnection to visitors/connection.ts — breaks node.ts→to-sql.ts import cycle; use ArelConnection directly in ToSqlCtor
